✨ Added the deploy code for backend using .env value
