+++ packages/user_support/src/core/constVals.ts
@@ -1,0 +1,10 @@
+/** プロジェクトのルートに置かれる.clasp.jsonのパス */
+export const CLASP_JSON_PATH = '../../.clasp.json';
+
+/** AWS用の.envファイルのパス */
+export const PROXY_ENV_PATH = '../proxy/.env.local';
+/** バックエンド用の.envファイルのパス */
+export const BACKEND_ENV_PATH = '../backend/.env.local';
+
+/** AWSにおいてデプロイされたスタックの結果を格納するファイルのパス */
+export const AWS_STACK_RESULT_PATH = '../proxy/stackResult.json';
